feat(create-plugin): use react.lazy to code split scenes-app templates
