+++ packages/create-plugin/templates/scenes-app/src/components/App/App.tsx
@@ -3,12 +3,10 @@ import { AppRootProps } from '@grafana/data';
 import { PluginPropsContext } from '../../utils/utils.plugin';
 import { Routes } from '../Routes';
 
-export class App extends React.PureComponent<AppRootProps> {
-  render() {
-    return (
-      <PluginPropsContext.Provider value={this.props}>
-        <Routes />
-      </PluginPropsContext.Provider>
-    );
-  }
+export function App(props: AppRootProps) {
+  return (
+    <PluginPropsContext.Provider value={this.props}>
+      <Routes />
+    </PluginPropsContext.Provider>
+  );
 }
