+++ packages/create-plugin/templates/scenes-app/src/module.tsx
@@ -1,10 +1,10 @@
+import { lazy } from 'react';
 import { AppPlugin } from '@grafana/data';
-import { App } from './components/App';
-import { AppConfig } from './components/AppConfig';
+import { App } from './components/App/App';
 
 export const plugin = new AppPlugin<{}>().setRootPage(App).addConfigPage({
   title: 'Configuration',
   icon: 'cog',
-  body: AppConfig,
+  body: lazy(() => import('./components/AppConfig/AppConfig')),
   id: 'configuration',
 });
