+++ packages/create-plugin/templates/scenes-app/src/pages/WithDrilldown/WithDrilldown.tsx
@@ -93,8 +93,10 @@ const getDrilldownsAppScene = () => {
   });
 };
 
-export const WithDrilldown = () => {
+const WithDrilldown = () => {
   const scene = useMemo(() => getDrilldownsAppScene(), []);
 
   return <scene.Component model={scene} />;
 };
+
+export default WithDrilldown;
