+++ packages/create-plugin/templates/scenes-app/src/pages/WithTabs/WithTabs.tsx
@@ -38,8 +38,10 @@ const getScene = () =>
     ],
   });
 
-export const PageWithTabs = () => {
+const PageWithTabs = () => {
   const scene = useMemo(() => getScene(), []);
 
   return <scene.Component model={scene} />;
 };
+
+export default PageWithTabs;
